insighflow-ai/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
aiosqlite>=0.20.0
email-validator>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pyarrow>=18.0.0
chardet>=5.0.0
pytest-cov>=6.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
python-json-logger>=2.0.0
python-dateutil>=2.8.0
